+++ src/parser/nodes/SubroutineDefinitionNode.ts
@@ -1,6 +1,7 @@
 import { AstVisitor } from '../AstVisitor';
 import { AstNode } from './AstNode';
 import { BlockStatementNode } from './BlockStatementNode';
+import { ExpressionNode } from './expressions';
 import { Range } from './Range';
 import { StatementNode } from './StatementNode';
 
@@ -11,6 +12,7 @@ export class SubroutineDefinitionNode extends BlockStatementNode {
     body: StatementNode[],
     readonly labelTokenRange: Range,
     readonly endTokenRange: Range,
+    readonly returnValue?: ExpressionNode,
     parent?: AstNode
   ) {
     super(range, body, parent);